Refactor return statement to use unsigned short

+++ dna.cxx
@@ -20,8 +20,7 @@ bool DNA::getCode(const int i) const {
      * bit i = 1 --> result is 1, true
      * bit i = 0 --< result is 0, false
     */
-    return (static_cast<unsigned int>(codes) >> i) & 1U;    // the unsigned short might work here too
-    //return (codes >> i) & 1U;    // using unsigned short
+    return (codes >> i) & 1U;    // using unsigned short
 }
 
 void DNA::setCode(const int i, bool v) {
